reorganize files and update functions firm
